Fix disabled function test issues
